fix: trigger point events immediately following session-start skipping content check
